backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.0.4
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.4.0
python-dotenv==0.21.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
